[Refactor] Move getNewId funciton to utils
