+++ src/guillo-gmi/actions/copy_items.js
@@ -2,19 +2,9 @@ import React from 'react'
 import { PathTree } from '../components/modal'
 import { useTraversal } from '../contexts'
 import { getContainerFromPath } from '../lib/client'
+import { getNewId } from '../lib/utils'
 
 const withError = (res) => res.status >= 300
-
-function getNewId(id = '') {
-  const suffix = '-copy-'
-  const rgx = new RegExp(`($|${suffix}\\d*)`)
-
-  return id.replace(rgx, (r) => {
-    const num = parseInt(r.replace(suffix, '') || '0')
-    return `${suffix}${num + 1}`
-  })
-}
-
 export function CopyItems(props) {
   const Ctx = useTraversal()
   const { items = [] } = props
