+++ src/guillo-gmi/lib/utils.js
@@ -5,3 +5,13 @@ export const formatDate = (str) => {
     d.getMonth() + 1
   }/${d.getFullYear()} ${d.getHours()}:${minutes}`
 }
+
+export function getNewId(id = '') {
+  const suffix = '-copy-'
+  const rgx = new RegExp(`($|${suffix}\\d*)`)
+
+  return id.replace(rgx, (r) => {
+    const num = parseInt(r.replace(suffix, '') || '0')
+    return `${suffix}${num + 1}`
+  })
+}
